move command to commands file
