+++ build-tools/packages/build-cli/src/library/commands/generateEntrypoints.ts
@@ -7,13 +7,10 @@ import fs from "node:fs/promises";
 import path from "node:path";
 
 import type { PackageJson } from "@fluidframework/build-tools";
-import { Flags } from "@oclif/core";
 import type { ExportSpecifierStructure, Node } from "ts-morph";
 import { ModuleKind, Project, ScriptKind } from "ts-morph";
 
 import type { CommandLogger } from "../../logging.js";
-import { BaseCommand } from "./base.js";
-
 import { ApiLevel } from "../apiLevel.js";
 import { ApiTag } from "../apiTag.js";
 import type { ExportData, Node10CompatExportData } from "../packageExports.js";
@@ -22,7 +19,7 @@ import { getApiExports, getPackageDocumentationText } from "../typescriptApi.js"
 
 import { unscopedPackageNameString } from "./constants.js";
 
-const optionDefaults = {
+export const optionDefaults = {
 	mainEntrypoint: "./src/index.ts",
 	outDir: "./lib",
 	outFilePrefix: "",
@@ -32,126 +29,6 @@ const optionDefaults = {
 	outFilePublic: ApiLevel.public,
 	outFileSuffix: ".d.ts",
 } as const;
-
-/**
- * Generates type declarations files for Fluid Framework APIs to support API levels (/alpha, /beta. etc.).
- */
-export class GenerateEntrypointsCommand extends BaseCommand<
-	typeof GenerateEntrypointsCommand
-> {
-	static readonly description =
-		`Generates type declaration entrypoints for Fluid Framework API levels (/alpha, /beta. etc.) as found in package.json "exports"`;
-
-	static readonly flags = {
-		mainEntrypoint: Flags.file({
-			description: "Main entrypoint file containing all untrimmed exports.",
-			default: optionDefaults.mainEntrypoint,
-			exists: true,
-		}),
-		outDir: Flags.directory({
-			description: "Directory to emit entrypoint declaration files.",
-			default: optionDefaults.outDir,
-			exists: true,
-		}),
-		outFilePrefix: Flags.string({
-			description: `File name prefix for emitting entrypoint declaration files. Pattern of '${unscopedPackageNameString}' within value will be replaced with the unscoped name of this package.`,
-			default: optionDefaults.outFilePrefix,
-		}),
-		outFileAlpha: Flags.string({
-			description: "Base file name for alpha entrypoint declaration files.",
-			default: optionDefaults.outFileAlpha,
-		}),
-		outFileBeta: Flags.string({
-			description: "Base file name for beta entrypoint declaration files.",
-			default: optionDefaults.outFileBeta,
-		}),
-		outFileLegacy: Flags.string({
-			description: "Base file name for legacy entrypoint declaration files.",
-			default: optionDefaults.outFileLegacy,
-		}),
-		outFilePublic: Flags.string({
-			description: "Base file name for public entrypoint declaration files.",
-			default: optionDefaults.outFilePublic,
-		}),
-		outFileSuffix: Flags.string({
-			description:
-				"File name suffix including extension for emitting entrypoint declaration files.",
-			default: optionDefaults.outFileSuffix,
-		}),
-		node10TypeCompat: Flags.boolean({
-			description: `Optional generation of Node10 resolution compatible type entrypoints matching others.`,
-		}),
-		...BaseCommand.flags,
-	};
-
-	public async run(): Promise<void> {
-		const { mainEntrypoint, node10TypeCompat } = this.flags;
-
-		const packageJson = await readPackageJson();
-
-		const {
-			mapQueryPathToApiTagLevel,
-			mapApiTagLevelToOutput,
-			mapNode10CompatExportPathToData,
-		} = getOutputConfiguration(this.flags, packageJson, this.logger);
-
-		const promises: Promise<void>[] = [];
-
-		// Requested specific outputs that are not in the output map are explicitly
-		// removed for clean incremental build support.
-		for (const [outputPath, apiLevel] of mapQueryPathToApiTagLevel.entries()) {
-			if (
-				apiLevel !== undefined &&
-				typeof outputPath === "string" &&
-				!mapApiTagLevelToOutput.has(apiLevel)
-			) {
-				promises.push(fs.rm(outputPath, { force: true }));
-			}
-		}
-
-		if (node10TypeCompat && mapNode10CompatExportPathToData.size === 0) {
-			throw new Error(
-				'There are no API level "exports" requiring Node10 type compatibility generation.',
-			);
-		}
-
-		if (mapApiTagLevelToOutput.size === 0) {
-			throw new Error(
-				`There are no package exports matching requested output entrypoints:\n\t${[
-					...mapQueryPathToApiTagLevel.keys(),
-				].join("\n\t")}`,
-			);
-		}
-
-		// In the past @alpha APIs could be mapped to /legacy via --outFileAlpha.
-		// When @alpha is mapped to /legacy, @beta should not be included in
-		// @alpha aka /legacy entrypoint.
-		const separateBetaFromAlpha = this.flags.outFileAlpha !== ApiLevel.alpha;
-		promises.push(
-			generateEntrypoints(
-				mainEntrypoint,
-				mapApiTagLevelToOutput,
-				this.logger,
-				separateBetaFromAlpha,
-			),
-		);
-
-		if (node10TypeCompat) {
-			promises.push(
-				generateNode10TypeEntrypoints(mapNode10CompatExportPathToData, this.logger),
-			);
-		}
-
-		// All of the output actions (deletes of stale files or writing of new/updated files)
-		// are all independent and can be done in parallel.
-		await Promise.all(promises);
-	}
-}
-
-async function readPackageJson(): Promise<PackageJson> {
-	const packageJson = await fs.readFile("./package.json", { encoding: "utf8" });
-	return JSON.parse(packageJson) as PackageJson;
-}
 
 /**
  * Returns the path "prefix" for all of the output files.
@@ -205,7 +82,7 @@ function getLocalUnscopedPackageName(packageJson: PackageJson): string {
 	return unscopedPackageName;
 }
 
-function getOutputConfiguration(
+export function getOutputConfiguration(
 	flags: Readonly<Record<keyof typeof optionDefaults, string>> & { node10TypeCompat: boolean },
 	packageJson: PackageJson,
 	logger?: CommandLogger,
@@ -331,7 +208,7 @@ const generatedHeader: string = `/*!
  * @param log - logger
  * @param separateBetaFromAlpha - if true, beta APIs will not be included in alpha outputs
  */
-async function generateEntrypoints(
+export async function generateEntrypoints(
 	mainEntrypoint: string,
 	mapApiTagLevelToOutput: Map<ApiTag, ExportData>,
 	log: CommandLogger,
@@ -461,7 +338,7 @@ async function generateEntrypoints(
 	await Promise.all(fileSavePromises);
 }
 
-async function generateNode10TypeEntrypoints(
+export async function generateNode10TypeEntrypoints(
 	mapExportPathToData: Map<string, Node10CompatExportData>,
 	log: CommandLogger,
 ): Promise<void> {
